backend/interview_backend/requirements.txt
# Core Django + REST
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.1

# LangChain + AI
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.7

# LLM client (depends on what you use)
openai>=1.0.0

# PDF / DOCX parsing
pypdf2>=3.0.1
python-docx>=1.1.0

# Data validation
pydantic>=2.8.0

# Environment variables
python-dotenv>=1.0.1

# Database (SQLite built-in, but include Postgres if needed)
psycopg2-binary>=2.9.9

# For async support
httpx>=0.27.0

# Testing
pytest>=8.2.0
pytest-django>=4.8.0
